requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
psycopg2-binary==2.9.4
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
